pyproject.toml



[project]
name = "dagnosis"

version = "0.1.0"

description = "DAGnosis: Localized Identification of Data Inconsistencies using Structures"

readme = "README.md"

requires-python = ">=3.10"

license = {file = "LICENSE"}

keywords = ["dag", "data-centric", "machine", "learning", "structure"]

authors = [
  {name = "Nicolas Huynh", email = "nicolas.huynh99@gmail.com" }
]

maintainers = [
  {name = "Nicolas Huynh", email = "nicolas.huynh99@gmail.com" }
]

classifiers = [
  "Development Status :: 3 - Alpha",

  "Intended Audience :: Science/Research",

  "Topic :: Scientific/Engineering :: Artificial Intelligence",

  "License :: OSI Approved :: MIT License",

  "Programming Language :: Python :: 3.10",
]

dependencies = [
   "dill==0.3.6",
   "hydra-core==1.3.2",
   "lightgbm==3.3.3",
   "mapie==0.5.0",
   "matplotlib==3.5.3",
   "networkx==2.6.3",
   "numpy==1.21.6",
   "omegaconf==2.3.0",
   "pandas==1.3.5",
   "python-igraph==0.10.2",
   "pytorch-lightning",
   "scikit-learn==1.0.2",
   "scipy==1.7.3",
   "torch==1.13.0",
   "tqdm==4.57.0",
]


[project.optional-dependencies]
test = [
  "pytest",
  "coverage",
  "mypy",
  "black",
  "pytest-cov",
  "pytest-mypy",
  "pytest-black",
  "pytest-isort",
  "pytest-flakes",
  "pre-commit",
  "types-PyYAML",
]

[project.urls]
"Homepage" = "https://github.com/nicolashuynh/DAGNOSIS"

[build-system]
# AVOID CHANGING REQUIRES: IT WILL BE UPDATED BY PYSCAFFOLD!
requires = ["setuptools>=46.1.0", "setuptools_scm[toml]>=5"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
# For smarter version schemes and other configuration options,
# check out https://github.com/pypa/setuptools_scm
version_scheme = "no-guess-dev"


[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra --flakes --isort --mypy --cov --cov-report html --cov-report term"
testpaths = [
    "tests",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
exclude = [
    '^file1\.py$',  # TOML literal string (single-quotes, no escaping necessary)
]
ignore_missing_imports = true

[tool.isort]
profile = "black"
